Move favorite button to toolbar top trailing edge
